+++ formula-rxjava3/src/main/java/com/instacart/formula/rxjava3/RxStream.kt
@@ -1,4 +1,9 @@
+@file:Suppress("unused")
+
 package com.instacart.formula.rxjava3
 
-@Deprecated("use RxAction")
+@Deprecated(
+    "use RxAction",
+    replaceWith = ReplaceWith("RxAction", "com.instacart.formula.rxjava3.RxAction")
+)
 typealias RxStream<Event> = RxAction<Event>