+++ formula/src/test/java/com/instacart/formula/subjects/OnlyUpdateFormula.kt
@@ -1,18 +1,18 @@
 package com.instacart.formula.subjects
 
+import com.instacart.formula.ActionBuilder
 import com.instacart.formula.Evaluation
 import com.instacart.formula.Snapshot
 import com.instacart.formula.StatelessFormula
-import com.instacart.formula.StreamBuilder
 
 class OnlyUpdateFormula<Input>(
-    private val build: StreamBuilder<*, Unit>.(Input) -> Unit
+    private val build: ActionBuilder<*, Unit>.(Input) -> Unit
 ) : StatelessFormula<Input, Unit>() {
 
     override fun Snapshot<Input, Unit>.evaluate(): Evaluation<Unit> {
         return Evaluation(
             output = Unit,
-            updates = context.updates {
+            actions = context.actions {
                 build(this, input)
             }
         )