+++ formula/src/test/java/com/instacart/formula/subjects/RemovingTerminateStreamSendsNoMessagesFormula.kt
@@ -1,9 +1,9 @@
 package com.instacart.formula.subjects
 
+import com.instacart.formula.Action
 import com.instacart.formula.Evaluation
 import com.instacart.formula.Snapshot
 import com.instacart.formula.StatelessFormula
-import com.instacart.formula.Stream
 
 class RemovingTerminateStreamSendsNoMessagesFormula : StatelessFormula<RemovingTerminateStreamSendsNoMessagesFormula.Input, Unit>() {
 
@@ -14,10 +14,10 @@ class RemovingTerminateStreamSendsNoMessagesFormula : StatelessFormula<RemovingT
     override fun Snapshot<Input, Unit>.evaluate(): Evaluation<Unit> {
         return Evaluation(
             output = Unit,
-            updates = context.updates {
+            actions = context.actions {
                 val onTerminate = input.onTerminate
                 if (onTerminate != null) {
-                    events(Stream.onTerminate()) {
+                    events(Action.onTerminate()) {
                         transition(onTerminate)
                     }
                 }