+++ formula/src/test/java/com/instacart/formula/subjects/StartStopFormula.kt
@@ -29,9 +29,9 @@ class StartStopFormula(runtime: TestableRuntime) : Formula<Unit, State, Output>(
 
     override fun Snapshot<Unit, State>.evaluate(): Evaluation<Output> {
         return Evaluation(
-            updates = context.updates {
+            actions = context.actions {
                 if (state.listenForEvents) {
-                    events(incrementEvents.stream()) {
+                    events(incrementEvents.action()) {
                         transition(state.copy(count = state.count + 1))
                     }
                 }