+++ formula/src/test/java/com/instacart/formula/subjects/StreamInitMessageDeliveredOnce.kt
@@ -1,9 +1,9 @@
 package com.instacart.formula.subjects
 
+import com.instacart.formula.Action
 import com.instacart.formula.Evaluation
 import com.instacart.formula.Snapshot
 import com.instacart.formula.StatelessFormula
-import com.instacart.formula.Stream
 import com.instacart.formula.test.TestableRuntime
 
 object StreamInitMessageDeliveredOnce {
@@ -16,8 +16,8 @@ object StreamInitMessageDeliveredOnce {
 
             return Evaluation(
                 output = Unit,
-                updates = context.updates {
-                    Stream.onInit().onEvent {
+                actions = context.actions {
+                    Action.onInit().onEvent {
                         transition { timesInitializedCalled += 1 }
                     }
                 }