+++ formula/src/test/java/com/instacart/formula/subjects/StreamInputFormula.kt
@@ -1,9 +1,9 @@
 package com.instacart.formula.subjects
 
+import com.instacart.formula.Action
 import com.instacart.formula.Evaluation
 import com.instacart.formula.Snapshot
 import com.instacart.formula.StatelessFormula
-import com.instacart.formula.Stream
 
 class StreamInputFormula : StatelessFormula<Int, Unit>()  {
     val messages = mutableListOf<Int>()
@@ -11,8 +11,8 @@ class StreamInputFormula : StatelessFormula<Int, Unit>()  {
     override fun Snapshot<Int, Unit>.evaluate(): Evaluation<Unit> {
         return Evaluation(
             output = Unit,
-            updates = context.updates {
-                Stream.onData(input).onEvent {
+            actions = context.actions {
+                Action.onData(input).onEvent {
                     transition { messages.add(it) }
                 }
             }